Checkpoint before launcher refactor
