+++ client/lib/router/router.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:tms_client/base/base_scaffold.dart';
-import 'package:tms_client/deferred_widget.dart';
 import 'package:tms_client/providers/auth_provider.dart';
+import 'package:tms_client/router/deferred_widget.dart';
 import 'package:tms_client/views/home/home_view.dart' deferred as home;
 import 'package:tms_client/views/login/login_view.dart' deferred as login;
 import 'package:tms_client/views/setup/setup_view.dart' deferred as setup;
