+++ client/web/index.html
@@ -17,8 +17,64 @@
 
     <title>TMS</title>
     <link rel="manifest" href="manifest.json" />
+
+    <style>
+      /* Remove default margins and set full height */
+      html,
+      body {
+        margin: 0;
+        padding: 0;
+        height: 100%;
+      }
+
+      /* Dark mode by default, light mode if user prefers */
+      body {
+        background-color: #20222f;
+      }
+
+      /* Switch to white background if user has light mode preference */
+      @media (prefers-color-scheme: light) {
+        body {
+          background-color: #ffffff;
+        }
+        #loading-text {
+          color: #000000 !important;
+        }
+      }
+
+      /* Center everything on screen */
+      #loading {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        height: 100%;
+      }
+
+      /* Style the icon */
+      #loading-icon {
+        width: 120px;
+        height: 120px;
+        margin-bottom: 20px;
+      }
+
+      /* Style the loading text */
+      #loading-text {
+        color: #ffffff;
+        font-family:
+          -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
+        font-size: 18px;
+        font-weight: 400;
+      }
+    </style>
   </head>
   <body>
+    <!-- This div will be replaced by Flutter when it loads -->
+    <div id="loading">
+      <img id="loading-icon" src="favicon.svg" alt="TMS" />
+      <div id="loading-text">Loading...</div>
+    </div>
+
     <script src="flutter_bootstrap.js" async></script>
   </body>
 </html>
