+++ launcher/Cargo.toml
@@ -9,7 +9,8 @@ path = "src/bin/main.rs"
 
 [dependencies]
 anyhow = "1.0.100"
-eframe = "0.33.3"
+eframe = { version = "0.33.3", features = ["wayland", "x11"] }
+
 egui = "0.33.3"
 egui_extras = { version = "0.33.3", features = ["all_loaders"] }
 image = { version = "0.25.9", features = ["png", "jpeg", "gif"] }
